calendar_agent/pyproject.toml
[project]
name = "calendar-agent-a2a"
version = "0.1.0"
description = "Calendar agent with A2A server and A2UI support"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk>=0.3.0",
    "google-adk>=1.8.0",
    "google-genai>=1.27.0",
    "python-dotenv>=1.1.0",
    "jsonschema>=4.0.0",
    "a2ui",
    "click>=8.1.8",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[[tool.uv.index]]
url = "https://pypi.org/simple"
default = true
